requirements.txt
pdfplumber
requests
python-dotenv
google-auth-oauthlib
google-auth-httplib2
google-api-python-client